+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/atomist/rug-editors/compare/0.25.0...HEAD
+[Unreleased]: https://github.com/atomist/rug-editors/compare/0.26.0...HEAD
+
+## [0.26.0] - 2017-04-22
+
+[0.26.0]: https://github.com/atomist/rug-editors/compare/0.25.0...0.26.0
+
+Update release
 
 ### Changed